Remove useless casts from code

+++ src/jpeg2000rct.rs
@@ -25,32 +25,27 @@ impl RCT<u16> for u8 {
         offset: usize,
         _bits: usize,
     ) {
-        let Y = &src[0][offset as usize..];
-        let Cb = &src[1][offset as usize..];
-        let Cr = &src[2][offset as usize..];
-        for y in 0..height as usize {
-            for x in 0..width as usize {
-                let Cbtmp = Cb[(y * stride as usize) + x] as i32 - (1 << 8); // Missing from spec
-                let Crtmp = Cr[(y * stride as usize) + x] as i32 - (1 << 8); // Missing from spec
-                let green = Y[(y * stride as usize) + x] as i32
-                    - ((Cbtmp as i32 + Crtmp as i32) >> 2);
-                let red = Crtmp as i32 + green;
-                let blue = Cbtmp as i32 + green;
-                dst[0][offset as usize + (y * stride as usize) + x] =
-                    green as u8;
-                dst[1][offset as usize + (y * stride as usize) + x] =
-                    blue as u8;
-                dst[2][offset as usize + (y * stride as usize) + x] =
-                    red as u8;
+        let Y = &src[0][offset..];
+        let Cb = &src[1][offset..];
+        let Cr = &src[2][offset..];
+        for y in 0..height {
+            for x in 0..width {
+                let Cbtmp = Cb[(y * stride) + x] - (1 << 8); // Missing from spec
+                let Crtmp = Cr[(y * stride) + x] - (1 << 8); // Missing from spec
+                let green = Y[(y * stride) + x] - ((Cbtmp + Crtmp) >> 2);
+                let red = Crtmp + green;
+                let blue = Cbtmp + green;
+                dst[0][offset + (y * stride) + x] = green as u8;
+                dst[1][offset + (y * stride) + x] = blue as u8;
+                dst[2][offset + (y * stride) + x] = red as u8;
             }
         }
         if src.len() == 4 {
-            let s = &src[3][offset as usize..];
-            let d = &mut dst[3][offset as usize..];
-            for y in 0..height as usize {
-                for x in 0..width as usize {
-                    d[(y * stride as usize) + x] =
-                        s[(y * stride as usize) + x] as u8;
+            let s = &src[3][offset..];
+            let d = &mut dst[3][offset..];
+            for y in 0..height {
+                for x in 0..width {
+                    d[(y * stride) + x] = s[(y * stride) + x] as u8;
                 }
             }
         }
@@ -71,29 +66,17 @@ impl RCT<u8> for u16 {
         bits: usize,
     ) {
         let src = dst;
-        for y in 0..height as usize {
-            for x in 0..width as usize {
-                let Cbtmp = (src[1]
-                    [offset as usize + (y * stride as usize) + x]
-                    as i32
-                    - 1)
-                    << bits as i32; // Missing from spec
-                let Crtmp = (src[2]
-                    [offset as usize + (y * stride as usize) + x]
-                    as i32
-                    - 1)
-                    << bits as i32; // Missing from spec
-                let blue = src[0][offset as usize + (y * stride as usize) + x]
-                    as i32
-                    - ((Cbtmp as i32 + Crtmp as i32) >> 2);
-                let red = Crtmp as i32 + blue;
-                let green = Cbtmp as i32 + blue;
-                src[0][offset as usize + (y * stride as usize) + x] =
-                    green as u16;
-                src[1][offset as usize + (y * stride as usize) + x] =
-                    blue as u16;
-                src[2][offset as usize + (y * stride as usize) + x] =
-                    red as u16;
+        for y in 0..height {
+            for x in 0..width {
+                let Cbtmp = (src[1][offset + (y * stride) + x] - 1) << bits; // Missing from spec
+                let Crtmp = (src[2][offset + (y * stride) + x] - 1) << bits; // Missing from spec
+                let blue =
+                    src[0][offset + (y * stride) + x] - ((Cbtmp + Crtmp) >> 2);
+                let red = Crtmp + blue;
+                let green = Cbtmp + blue;
+                src[0][offset + (y * stride) + x] = green as u16;
+                src[1][offset + (y * stride) + x] = blue as u16;
+                src[2][offset + (y * stride) + x] = red as u16;
             }
         }
     }
@@ -112,32 +95,27 @@ impl RCT<u32> for u16 {
         offset: usize,
         _bits: usize,
     ) {
-        let Y = &src[0][offset as usize..];
-        let Cb = &src[1][offset as usize..];
-        let Cr = &src[2][offset as usize..];
-        for y in 0..height as usize {
-            for x in 0..width as usize {
-                let Cbtmp = Cb[(y * stride as usize) + x] as i32 - (1 << 16); // Missing from spec
-                let Crtmp = Cr[(y * stride as usize) + x] as i32 - (1 << 16); // Missing from spec
-                let green = Y[(y * stride as usize) + x] as i32
-                    - ((Cbtmp as i32 + Crtmp as i32) >> 2);
-                let red = Crtmp as i32 + green;
-                let blue = Cbtmp as i32 + green;
-                dst[0][offset as usize + (y * stride as usize) + x] =
-                    green as u16;
-                dst[1][offset as usize + (y * stride as usize) + x] =
-                    blue as u16;
-                dst[2][offset as usize + (y * stride as usize) + x] =
-                    red as u16;
+        let Y = &src[0][offset..];
+        let Cb = &src[1][offset..];
+        let Cr = &src[2][offset..];
+        for y in 0..height {
+            for x in 0..width {
+                let Cbtmp = Cb[(y * stride) + x] - (1 << 16); // Missing from spec
+                let Crtmp = Cr[(y * stride) + x] - (1 << 16); // Missing from spec
+                let green = Y[(y * stride) + x] - ((Cbtmp + Crtmp) >> 2);
+                let red = Crtmp + green;
+                let blue = Cbtmp + green;
+                dst[0][offset + (y * stride) + x] = green as u16;
+                dst[1][offset + (y * stride) + x] = blue as u16;
+                dst[2][offset + (y * stride) + x] = red as u16;
             }
         }
         if src.len() == 4 {
-            let s = &src[3][offset as usize..];
-            let d = &mut dst[3][offset as usize..];
-            for y in 0..height as usize {
-                for x in 0..width as usize {
-                    d[(y * stride as usize) + x] =
-                        s[(y * stride as usize) + x] as u16;
+            let s = &src[3][offset..];
+            let d = &mut dst[3][offset..];
+            for y in 0..height {
+                for x in 0..width {
+                    d[(y * stride) + x] = s[(y * stride) + x] as u16;
                 }
             }
         }
